chore: refactor scim, get rid of hooks and move to separate tab
